-Yshow-symkinds: prints kinds next to symbol names
This very situational option proved to be very useful when debugging
https://issues.scala-lang.org/browse/SI-5415

With the help of -Yshow-symkinds, it became possible to distinguish
a free var from a module symbol, which gave precise indication of the
root of the bug.

This changeset also upgrades stringification of symbols and types, so
I'd like to get a review by @paulp and @odersky.