Convert values to Int in switchable patterns.
Further improvements to how -Yvirtpatmat
handles switch-like patterns that can be
translated to switch tables.

First of all, we added a check whether
a type of an expression we pattern match
on is in the set of allowed types for
switch patterns.

If yes, we translate a pattern to switch
one by converting both an expression we
pattern match on and literals in a pattern
to an Int.

I borrowed an idea of converting to Ints
from both old pattern matcher implementation
and from how javac handles it.