Added a test-case for switches with Yvirtpatmat.
Added a bunch of tests that cover changes related
to switches that were applied to Yvirtpatmat
implementation.

Note: I didn't add those tests progressively because
my changes fix trees after typer phase but do not
affect resulting bytecode. How come? It's because
-Yvirtpatmat will emit pattern for switches and then
the old pattern matcher implementation would transform
them in the old fashion in explicitouter. We cannot
disable the old pattern matcher in explicitouter yet
because it doesn't handle patterns used for catching
exceptions.

Thus, consider this as a sign of the fact that Yvirtpatmat
is still work in progress.