Guard List_apply from premature forcitude.
Solved the annotation cycle puzzle.  Was being burned again by
the hack which preferentially treats List() as Nil to avoid List's
extractor.

This commit includes a new starr which fully bootstraps.

Also at this point traits can extend Any and will not be given
an AnyRef parent (until erasure.) This is the case for AnyVal
and NotNull.