Fixes https://issues.scala-lang.org/browse/SI-5334
New version of reification isn't susceptible to this bug.

The problem was with Code.lift generating not only a tree, but also
a manifest with the type of that tree. That led to an issue in the case
of the type of the manifest depending on a class declared inside the
quasiquote.

Now manifests in reification are gone, so is the problem.