Remove alphanumeric characters from the major part of the Scala version number.
In practice, this removes the 'v' in v2.10.0-M1-0168-g99844ebc10-2012-02-08. Why is this important, you ask?
Many tools assume the version number to be composed of 3 numbers and a string qualifier. Eclipse is one of the
tools that validates version numbers, and this string is used to stamp Scala IDE builds against Scala nightlies.