Another existential problem down.
There is a window of danger when multiple related elements are
being typed where something which is conceptually one thing can
slip into two things, and those two things can be incompatible
with one another.  Less mysteriously, c478eb770d fixed this:

  def f = { object Bob ; Bob } ; val g = f

But, it did not fix this:

  def f = { case class Bob() ; Bob } ; val g = f

See test case pos/existentials-harmful.scala for an "in the wild"
code example fixed by this commit.

The root of the problem was that the getter and the field would each
independently derive the same existential type to describe Bob, but
those existentials were not the same as one another.

This has been the most elusive bug I have ever fixed. I want to cry when
I think of how much time I've put into it over the past half decade or
so. Unfortunately the way the repl works it is particularly good at
eliciting those grotesque found/required error messages and so I was
never able to let the thing go.

There is still a cosmetic issue (from the last commit really) where
compound types wind up with repeated parents.

Closes SI-1195, SI-1201.