Fixed bug in DefDef creation.
What's improbable about 29f933b60 is that a good chunk of the commit
exists only so you can pass "mods" to the DefDef constructor. And then
the body isn't updated to use it, the parameter dies on the vine. That
was five years ago.

I logged when mods didn't match Modifiers(sym.flags) and there were
a bunch of differences, but I didn't see any interesting behavior
emerging. Still, I bet something changed somewhere (hopefully in the
fixy direction.)