Updated existentials test with non-working parts.
Lazy accessors have the same essential problem as described in
recent commit messages, so this still doesn't work:

  def f() = { case class Bob(); Bob } ; lazy val g = f

On the whole I'm trying to solve this at the wrong level. The derived
accessor methods and fields of a declaration should not ever wander far
enough apart that there is any challenge in reconciling them. (The same
is true for case classes/objects.) They're dependent synthetics who know
about one another from the beginning, all we have to do is not forget.
In the meantime, test case.