Fix for SI-5452.
Why is calling the error function not enough to register the error, why
does "setError(tree)" have to be called as well? That was the cause of
this particular stackoverflow. In ContextErrors I see lots of methods
which call setError and lots more which do not, and frankly it's all
pretty terrifying.

There is zero documentation attached to setError.  Maybe there's an
explanation somewhere I'm not seeing.

Review by @hubertp.