Added test files to verify previous commit.
Tests scalac -optimize -Xprint:specialize -Ylog:inliner output to
verify that final/@inline + specialization are being handled
correctly (that is, the original class' specialized methods should
not be final/@inline, but its specialized subclass' should be).

This test was written by Vlad Ureche based on the bug report
in SI-5005.