Fix for "type _$1 defined twice".
Actually not a fix, only a workaround.  Package objects strike
again: you don't want to have existentials in there, like this:

  private val resolverFunction: PartialFunction[Throwable, Try[_]] = ...

There are a few irrelevant-to-that-bug changes included which I
did while debugging and figured I'd keep them.  With this patch I
can merge/am merging the scala.concurrent work.