-Xplugin value passed by the Eclipse IDE are incorrectly parsed when it contains whitespaces.
Assume -Xplugin is given the value

  C:\Programs Files\plugins\Aplugin.jar C:\Programs Files\plugins\Bplugin.jar

Calling ``tryToSetFromPropertyValue`` with the above value will always result
in a total mess, no matter what, because it will split the string at
whitespaces.

The proposed solution is to change the implementation of
``tryToSetFromPropertyValue`` to use `,` (comma) as the splitting character

Further, I'm quite convinced that the current implementation of
``MultiStringSetting.tryToSetFromPropertyValue`` has never worked, that is why I
did not create an overload of ``tryToSetFromPropertyValue`` where the splitting
character (or string) can be passed as argument.

There is also an Eclipse Scala IDE associated to this issue:

  http://scala-ide-portfolio.assembla.com/spaces/scala-ide/tickets/1000917