Added updated ForkJoinPool, w/ necessary updates to Scala Actors.
This commit includes Doug Lea's updates to ForkJoinPool, as of January
25th, 2012. Details of the changes and performance improvements
available at: http://markmail.org/message/323vxzn6irkk5yrg.
The ForkJoinPool used in this commit comes from the most recent JSR166y.

Additionally, also included are minimal changes to parts of the Scala
Actors library which interface with the ForkJoinPool, as the
ForkJoinPool's interface has changed (prior to the release of Java 7)
since we last updated it for the Scala 2.8 release.

Of note- this is part of the planned overhaul of scala.concurrent,
and corresponds to ticket SI-5523.

For testing this was built on JDK 1.6, and passes all tests on both
JDK 1.5 and 1.6.

A new forkjoin.jar is necessary prior to applying these changes.
Using this source, the new jar can be built by running:

    ant newforkjoin forkjoin.done

This creates a new forkjoin.jar in build/libs/. It must replace
lib/forkjoin.jar.