Finally did something about broken irrefutability.
The parser has always been confused about tuple patterns in
for comprehensions.  It thinks it can fail to recognize an
irrefutable pattern and have it removed in refchecks, but it
is sadly mistaken, because the unnecessary filter has a tendency
to fail the compile in typer.

Look more intently for irrefutable patterns and don't insert
the unnecessary filter.  Closes SI-5589, SI-1336.