Fix memory leak related to default arguments in presentation compiler.
This bug exists for a long time, but was triggered / discovered only lately
by the default argument of mkParams in 118aef558f.

This reverts the workaroud commit 19b6ad5ee4.

The fix is tested by test/files/presentation/memory-leaks which runs the
presentation compiler several times on Typers.scala. I could not reproduce
the memory leak in a smaller test case.