Fix for range positions.
You are a fireman, leading a team of firemen on a rescue
in a burning building.  When you have cleared a room, you place
a sticker on the door which says "Room cleared!" If the other
firemen see that sticker, they will skip that room, knowing
there are no small children or adorable puppies behind it.

As the lead fireman, here is what you should not do: run from
door to door slapping "Room cleared!" stickers on them all, so
the other firemen shrug and go home.

Translation: when recursion depends on a condition, if you
alter the condition before you recurse, you may not recurse at all.