Trying to set an example in the land of macros.
Hopefully there are useful examples to be had in here.
Please note that you should never be testing "x.size != 0",
use isEmpty or nonEmpty.  Even if these happen to be small
collections, it's not a habit you want to form: size is O(n)
on a number of collections.  (Most pointedly, Lists.)