SI-5760: Improve error message for package$Klass conflict with Klass
Added a clarification to DoubleDefError for when the previous symbol
was in the package object but current symbol is not.

This was actually supposed to be an opportunity to hack partest to
run the two-step failing compilation, but somebody beat me to it and my
rebase failed.

The next hacking opportunity might be to add .pt script files!  The
possibilities are endless.