Restored portion of code removed in a7f68ce32c .
A hole in the net discovered by retronym makes me put back the
former special casing of module class type inference; there's no
reason to get it wrong.  It appears that

  TypeRef(pre1, sym1, Nil)
  SingleType(pre2, sym2)

are not seen as equivalent if they have different prefixes even
if the prefixes should be equivalent.  For now I'm leaving it; in
principle this patch should eventually be reverted after subtyping
and/or the representation of singleton types advances to the point
that it isn't necessary.  Closes SI-5777.