Clutch modification to tree printing.
Don't print trees under -Xprint:all if they're identical to the
tree printed at the previous phase.  It only works for a single
compilation unit but that is a huge step forward for us debuggers.
For instance this file:

  trait Foo { def f = 5 }

used to produce 332 lines of output and now produces 92, with
zero loss of information.  It ends with:

  [[syntax trees at end of                   cleanup]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of                     icode]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of                   inliner]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of   inlineExceptionHandlers]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of                  closelim]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of                       dce]] // a.scala: tree is unchanged since mixin
  [[syntax trees at end of                       jvm]] // a.scala: tree is unchanged since mixin