Specialize lazy vals (closes SI-5552)
Previously, specialized lazy vals would not work at all when used in
specialized classes, and would just return an uninitialized value.
After this patch, they work in the same way as other specialized fields
do (i.e. a new specialized field is created, and the specialized class
uses that instead of the base class' field).

Note that there are still known problems with specialized lazy vals
(for instance SI-4717) but it seemed to me that this was better than
nothing.