Widen types in names/defaults transformations.
We were getting away with this somehow, but the types are wrong after
typer and that sort of thing is noticed by more people now. I took the
opportunity to add our first -Ycheck:all test, which is at least as much
about helping -Ycheck:all remain in good working order as it is about
this test.