Closes SI-5821.
This was an interesting one. Basically an erroneous
import was creating an erroneous symbol for Array (similary
for other symbols that were 'found' in this import) which was
leading to all sorts of inconsistencies and spurious errors.

This wasn't a bug in ContextErrors but rather something that
existed for ages and was hidden from the general audience.

Review by @paulp.